refactor(constants): split oauth into oauth-clients and oauth-seed
The single ./oauth export forced every consumer to pull the seed layer's
@better-auth/oauth-provider types and APPS graph just to read a client id.
Split by dependency footprint:

- oauth-clients.ts: public client ids and scopes, zero non-type deps. Safe
  for a framework-agnostic client like @epicenter/auth to import.
- oauth-seed.ts: TrustedOAuthClient, buildTrustedOAuthClients, and
  projectTrustedOAuthClientToRow, the server-only trusted-client builders.

Consumers now declare which half they need; the published auth client can no
longer reach the seed logic.
