+++ apps/tab-manager/src/lib/platform/auth/auth.ts
@@ -10,7 +10,7 @@
 
 import { loadPersistedAuthStorage } from '@epicenter/auth';
 import { createExtensionOAuthLauncher } from '@epicenter/auth/oauth-launchers';
-import { EPICENTER_TAB_MANAGER_OAUTH_CLIENT_ID } from '@epicenter/constants/oauth';
+import { EPICENTER_TAB_MANAGER_OAUTH_CLIENT_ID } from '@epicenter/constants/oauth-clients';
 import { APP_URLS } from '@epicenter/constants/vite';
 import { storage } from '@wxt-dev/storage';
