+++ packages/constants/src/oauth-clients.ts
@@ -1,0 +1,39 @@
+/**
+ * Public OAuth client ids and scopes every Epicenter first-party app presents
+ * during sign-in.
+ *
+ * These are public PKCE client ids, not secrets: each identifies an app type,
+ * not a user, machine, install, or credential, and every install of a given
+ * app uses the same value. They are split out from the server-only trusted
+ * client builders (see oauth-seed.ts) so a framework-agnostic client like
+ * `@epicenter/auth` can import an id and the scopes without reaching the seed
+ * layer or its `@better-auth/oauth-provider` types.
+ */
+
+/**
+ * OAuth public client id for `epicenter auth login`.
+ *
+ * The CLI uses an out-of-band (OOB) authorization-code + PKCE flow against
+ * the same `/auth/oauth2/token` endpoint the browser uses. After sign-in
+ * on the hosted portal, Better Auth redirects to the API origin's
+ * `/auth/cli-callback`, which renders the one-time code; the user pastes
+ * it into the terminal. This identifies the CLI app type, not a user,
+ * machine, install, or secret. Every CLI install uses the same value.
+ */
+export const EPICENTER_CLI_OAUTH_CLIENT_ID = 'epicenter-cli';
+export const EPICENTER_FUJI_OAUTH_CLIENT_ID = 'epicenter-fuji';
+export const EPICENTER_FUJI_TAURI_OAUTH_REDIRECT_URI =
+	'epicenter-fuji://auth/callback';
+export const EPICENTER_HONEYCRISP_OAUTH_CLIENT_ID = 'epicenter-honeycrisp';
+export const EPICENTER_OPENSIDIAN_OAUTH_CLIENT_ID = 'epicenter-opensidian';
+export const EPICENTER_TAB_MANAGER_OAUTH_CLIENT_ID = 'epicenter-tab-manager';
+export const EPICENTER_ZHONGWEN_OAUTH_CLIENT_ID = 'epicenter-zhongwen';
+
+export const EPICENTER_OAUTH_SCOPES = [
+	'openid',
+	'profile',
+	'email',
+	'offline_access',
+] as const;
+
+export const EPICENTER_OAUTH_SCOPE = EPICENTER_OAUTH_SCOPES.join(' ');
