+++ packages/constants/src/oauth-seed.ts
@@ -1,5 +1,15 @@
 import type { SchemaClient } from '@better-auth/oauth-provider';
 import { APPS, appOrigins } from '#apps';
+import {
+	EPICENTER_CLI_OAUTH_CLIENT_ID,
+	EPICENTER_FUJI_OAUTH_CLIENT_ID,
+	EPICENTER_FUJI_TAURI_OAUTH_REDIRECT_URI,
+	EPICENTER_HONEYCRISP_OAUTH_CLIENT_ID,
+	EPICENTER_OAUTH_SCOPES,
+	EPICENTER_OPENSIDIAN_OAUTH_CLIENT_ID,
+	EPICENTER_TAB_MANAGER_OAUTH_CLIENT_ID,
+	EPICENTER_ZHONGWEN_OAUTH_CLIENT_ID,
+} from './oauth-clients.js';
 import { OAUTH_ROUTES } from './oauth-routes.js';
 
 /**
@@ -29,34 +39,6 @@ export type TrustedOAuthClient = {
 	>;
 	redirectUris: readonly string[];
 };
-
-/**
- * OAuth public client id for `epicenter auth login`.
- *
- * The CLI uses an out-of-band (OOB) authorization-code + PKCE flow against
- * the same `/auth/oauth2/token` endpoint the browser uses. After sign-in
- * on the hosted portal, Better Auth redirects to the API origin's
- * `/auth/cli-callback`, which renders the one-time code; the user pastes
- * it into the terminal. This identifies the CLI app type, not a user,
- * machine, install, or secret. Every CLI install uses the same value.
- */
-export const EPICENTER_CLI_OAUTH_CLIENT_ID = 'epicenter-cli';
-export const EPICENTER_FUJI_OAUTH_CLIENT_ID = 'epicenter-fuji';
-export const EPICENTER_FUJI_TAURI_OAUTH_REDIRECT_URI =
-	'epicenter-fuji://auth/callback';
-export const EPICENTER_HONEYCRISP_OAUTH_CLIENT_ID = 'epicenter-honeycrisp';
-export const EPICENTER_OPENSIDIAN_OAUTH_CLIENT_ID = 'epicenter-opensidian';
-export const EPICENTER_TAB_MANAGER_OAUTH_CLIENT_ID = 'epicenter-tab-manager';
-export const EPICENTER_ZHONGWEN_OAUTH_CLIENT_ID = 'epicenter-zhongwen';
-
-export const EPICENTER_OAUTH_SCOPES = [
-	'openid',
-	'profile',
-	'email',
-	'offline_access',
-] as const;
-
-export const EPICENTER_OAUTH_SCOPE = EPICENTER_OAUTH_SCOPES.join(' ');
 
 /**
  * Path every first-party app receives the OAuth callback at, on each of its
