+++ packages/server/src/auth/plugins.ts
@@ -1,9 +1,7 @@
 import { oauthProvider } from '@better-auth/oauth-provider';
 import { JWT_SIGNING_ALG } from '@epicenter/constants/auth';
-import {
-	buildTrustedOAuthClients,
-	EPICENTER_OAUTH_SCOPES,
-} from '@epicenter/constants/oauth';
+import { EPICENTER_OAUTH_SCOPES } from '@epicenter/constants/oauth-clients';
+import { buildTrustedOAuthClients } from '@epicenter/constants/oauth-seed';
 import type { BetterAuthOptions } from 'better-auth';
 import { jwt } from 'better-auth/plugins/jwt';
