+++ packages/server/src/middleware/require-auth.test.ts
@@ -18,7 +18,7 @@
 import { expect, test } from 'bun:test';
 import { oauthProvider } from '@better-auth/oauth-provider';
 import { JWT_SIGNING_ALG } from '@epicenter/constants/auth';
-import { EPICENTER_OAUTH_SCOPES } from '@epicenter/constants/oauth';
+import { EPICENTER_OAUTH_SCOPES } from '@epicenter/constants/oauth-clients';
 import { betterAuth } from 'better-auth';
 import { type MemoryDB, memoryAdapter } from 'better-auth/adapters/memory';
 import { jwt } from 'better-auth/plugins';
